스케줄 관리 쪽 수정 중

+++ SpringProject/src/main/java/com/dita/controller/AdminMemberApiController.java
@@ -3,9 +3,12 @@ package com.dita.controller;
 import com.dita.domain.Sched;
 import com.dita.domain.Type;
 import com.dita.persistence.AdminMemberRepository;
+import com.dita.persistence.UserRepository;
 import com.dita.vo.SchedDto;
 
 import lombok.RequiredArgsConstructor;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.DayOfWeek;
@@ -17,7 +20,9 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @RequestMapping("/api/admin/member")
 public class AdminMemberApiController {
-
+	
+	@Autowired
+	private UserRepository userRepository;
     private final AdminMemberRepository adminMemberRepository;
 
     @PostMapping("/update")
@@ -39,9 +44,15 @@ public class AdminMemberApiController {
     @DeleteMapping("/delete/{id}")
     public String deleteSchedule(@PathVariable int id) {
         if (adminMemberRepository.existsById(id)) {
-            adminMemberRepository.deleteById(id);
-            return "삭제 완료";
+        	Sched sched = adminMemberRepository.findById(id).orElse(null);
+            if (sched != null) {
+                String userId = sched.getUser().getUsersId();  // 사용자 ID 가져오기
+                adminMemberRepository.deleteById(id);         // 스케줄 삭제
+                userRepository.deleteById(userId);            // 사용자 삭제
+                return "삭제 완료";
+            }
         }
         return "삭제 실패: 해당 ID 없음";
     }
+
 }
